split the endpoints into separate files. Some of it is overkill but it at least demonstrates how to keep endpoints neat and clean.
